chore: move parseIni to it's own file
