+++ packages/shared-ini-file-loader/src/types.ts
@@ -1,0 +1,12 @@
+export interface Profile {
+  [key: string]: string | undefined;
+}
+
+export interface ParsedIniData {
+  [key: string]: Profile;
+}
+
+export interface SharedConfigFiles {
+  credentialsFile: ParsedIniData;
+  configFile: ParsedIniData;
+}
